+++ openframe-client-core/src/main/java/com/openframe/client/scheduler/ScheduleJobExecutionWatchdogScheduler.java
@@ -1,6 +1,6 @@
 package com.openframe.client.scheduler;
 
-import com.openframe.client.service.rmm.ScheduleExecutionHeaderWatchdogService;
+import com.openframe.client.service.rmm.watchdog.ScheduleJobExecutionWatchdogService;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnProperty;
@@ -11,15 +11,15 @@ import org.springframework.stereotype.Component;
 @RequiredArgsConstructor
 @Slf4j
 @ConditionalOnProperty(name = "openframe.rmm.schedule.watchdog.enabled", havingValue = "true")
-public class ScheduleExecutionHeaderWatchdogScheduler {
+public class ScheduleJobExecutionWatchdogScheduler {
 
-    private final ScheduleExecutionHeaderWatchdogService headerWatchdogService;
+    private final ScheduleJobExecutionWatchdogService watchdogService;
 
     @Scheduled(fixedDelayString = "${openframe.rmm.schedule.watchdog.interval:60000}")
     public void sweep() {
         log.debug("Running schedule-header watchdog sweep");
         try {
-            headerWatchdogService.markStuckSheduleJobsAsFailed();
+            watchdogService.markStuckScheduleJobsAsFailed();
         } catch (Exception e) {
             log.error("Schedule-header watchdog sweep failed", e);
         }
