+++ openframe-client-core/src/main/java/com/openframe/client/scheduler/ScriptExecutionWatchdogScheduler.java
@@ -1,24 +1,12 @@
 package com.openframe.client.scheduler;
 
-import com.openframe.client.service.rmm.ScriptExecutionWatchdogService;
+import com.openframe.client.service.rmm.watchdog.ScriptExecutionWatchdogService;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnProperty;
 import org.springframework.scheduling.annotation.Scheduled;
 import org.springframework.stereotype.Component;
 
-/**
- * Periodic backstop for Execution rows stuck in {@code RUNNING}. Delegates to
- * {@link ScriptExecutionWatchdogService}; this class is the cron wiring only.
- *
- * <p>Disabled by default — explicitly enable per environment via
- * {@code openframe.rmm.execution.watchdog.enabled=true}. The interval is
- * tuneable via {@code openframe.rmm.execution.watchdog.interval} (millis,
- * default 60s). The stuck-threshold is per-execution (each script's own
- * {@code timeoutSeconds} plus {@code openframe.rmm.execution.watchdog.grace-seconds},
- * default 120); rows without a timeout use the fixed
- * {@code openframe.rmm.execution.watchdog.threshold-seconds} (default 600).
- */
 @Component
 @RequiredArgsConstructor
 @Slf4j
@@ -31,7 +19,7 @@ public class ScriptExecutionWatchdogScheduler {
     public void sweep() {
         log.debug("Running Execution watchdog sweep");
         try {
-            watchdogService.markStuckExecutionsAsFailing();
+            watchdogService.markStuckExecutionsAsFailed();
         } catch (Exception e) {
             log.error("Execution watchdog sweep failed", e);
         }
