+++ openframe-client-core/src/main/java/com/openframe/client/service/rmm/watchdog/ScheduleJobExecutionWatchdogService.java
@@ -1,4 +1,4 @@
-package com.openframe.client.service.rmm;
+package com.openframe.client.service.rmm.watchdog;
 
 import com.openframe.data.document.rmm.ExecutionStatus;
 import com.openframe.data.document.rmm.ScheduleScriptExecution;
@@ -16,16 +16,11 @@ import java.util.List;
 @Service
 @RequiredArgsConstructor
 @Slf4j
-public class ScheduleExecutionHeaderWatchdogService {
+public class ScheduleJobExecutionWatchdogService {
 
     private final ScriptExecutionRepository scriptExecutionRepository;
     private final ScheduleScriptExecutionRepository scheduleScriptExecutionRepository;
 
-    /**
-     * How long a header may sit in RUNNING before the sweep considers it stuck. Should exceed the
-     * leaf watchdog's threshold so leaves are reaped first; the finalize is guarded regardless (a
-     * header with any still-RUNNING leaf is skipped), so this only bounds how eagerly the sweep looks.
-     */
     @Value("${openframe.rmm.schedule.watchdog.threshold-seconds:900}")
     private long thresholdSeconds;
 
@@ -35,29 +30,25 @@ public class ScheduleExecutionHeaderWatchdogService {
         }
         LeafStatusCounts counts = scriptExecutionRepository.countLeavesByStatus(tenantId, executionId);
         if (counts.running() > 0) {
-            return false;   // still in flight → header stays RUNNING
+            return false;// still in flight → header stays RUNNING
         }
         ExecutionStatus finalStatus;
         if (counts.failed() > 0) {
             finalStatus = ExecutionStatus.FAILED;
         } else {
-            // No running, no failed: all leaves succeeded, OR there are no leaves at all. A header
-            // with zero leaves never actually ran → FAILED, not a misleading SUCCESS.
-            boolean hasAnyLeaf = scriptExecutionRepository
-                    .findFirstByTenantIdAndExecutionId(tenantId, executionId).isPresent();
+            boolean hasAnyLeaf = scriptExecutionRepository.findFirstByTenantIdAndExecutionId(tenantId, executionId).isPresent();
             finalStatus = hasAnyLeaf ? ExecutionStatus.SUCCESS : ExecutionStatus.FAILED;
         }
-        long modified = scheduleScriptExecutionRepository.transitionIfRunning(
-                tenantId, executionId, finalStatus, Instant.now());
+
+        long modified = scheduleScriptExecutionRepository.transitionIfRunning(tenantId, executionId, finalStatus, Instant.now());
         if (modified > 0) {
-            log.info("Finalized schedule fire header: executionId={} status=RUNNING→{} tenantId={}",
-                    executionId, finalStatus, tenantId);
+            log.info("Finalized schedule fire header: executionId={} status=RUNNING→{} tenantId={}", executionId, finalStatus, tenantId);
         }
+
         return modified > 0;
     }
 
-    /** Sweep: finalize any RUNNING header older than the threshold whose leaves have all settled. */
-    public void markStuckSheduleJobsAsFailed() {
+    public void markStuckScheduleJobsAsFailed() {
         Instant now = Instant.now();
         List<ScheduleScriptExecution> candidates = scheduleScriptExecutionRepository
                 .findByStatusAndDispatchedAtBefore(ExecutionStatus.RUNNING, now.minusSeconds(thresholdSeconds));
