+++ openframe-data-mongo-common/src/main/java/com/openframe/data/document/rmm/ScheduleScriptExecution.java
@@ -27,7 +27,6 @@ import java.time.Instant;
         name = "tenant_schedule_dispatchedAt",
         def = "{'tenantId': 1, 'scheduleId': 1, 'dispatchedAt': -1}"
 )
-// Serves the header watchdog sweep (status = RUNNING AND dispatchedAt < threshold).
 @CompoundIndex(name = "status_dispatchedAt", def = "{'status': 1, 'dispatchedAt': 1}")
 public class ScheduleScriptExecution implements TenantScoped {
